some api changes; enabled ogg vorbis plugin
git-svn-id: svn://svn.code.sf.net/p/qmmp-dev/code/trunk/qmmp@565 90c681e8-e032-0410-971d-27865f9a5e38

+++ src/qmmp/decoder.h
@@ -41,7 +41,7 @@ public:
     // Standard Decoder API
     virtual bool initialize() = 0;
     virtual qint64 lengthInSeconds() = 0;
-    virtual void seek(double) = 0;
+    virtual void seek(qint64) = 0;
     virtual void stop() = 0;
     //virtual void pause() = 0;
     //virtual void stop() = 0;
